refacto: put all of the POI logic inside the POI context instead of the UI components, then created a custom hook for position
