+++ src/hooks/usePosition.jsx
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { PositionContext } from "../Contexts/PositionProvider";
+
+export function usePosition() {
+  return useContext(PositionContext);
+}
